feat: remove metadata from db if account deleted
