+++ lib/app/features/chat/e2ee/views/pages/one_to_one_messages_page.dart
@@ -5,7 +5,6 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:ion/app/components/skeleton/container_skeleton.dart';
 import 'package:ion/app/exceptions/exceptions.dart';
 import 'package:ion/app/extensions/extensions.dart';
 import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
@@ -16,19 +15,14 @@ import 'package:ion/app/features/chat/e2ee/views/components/one_to_one_messages_
 import 'package:ion/app/features/chat/model/database/chat_database.c.dart';
 import 'package:ion/app/features/chat/providers/conversation_messages_provider.c.dart';
 import 'package:ion/app/features/chat/providers/exist_chat_conversation_id_provider.c.dart';
-import 'package:ion/app/features/chat/providers/muted_conversations_provider.c.dart';
 import 'package:ion/app/features/chat/recent_chats/providers/selected_edit_message_provider.c.dart';
 import 'package:ion/app/features/chat/recent_chats/providers/selected_reply_message_provider.c.dart';
 import 'package:ion/app/features/chat/views/components/message_items/edit_message_info/edit_message_info.dart';
 import 'package:ion/app/features/chat/views/components/message_items/messaging_bottom_bar/messaging_bottom_bar.dart';
 import 'package:ion/app/features/chat/views/components/message_items/replied_message_info/replied_message_info.dart';
-import 'package:ion/app/features/user_metadata/providers/user_metadata_from_db_provider.c.dart';
 import 'package:ion/app/features/user_metadata/providers/user_metadata_sync_provider.c.dart';
 import 'package:ion/app/hooks/use_on_init.dart';
-import 'package:ion/app/router/app_routes.c.dart';
 import 'package:ion/app/services/media_service/media_service.c.dart';
-import 'package:ion/app/utils/username.dart';
-import 'package:ion/generated/assets.gen.dart';
 
 class OneToOneMessagesPage extends HookConsumerWidget {
   const OneToOneMessagesPage({
@@ -124,88 +118,9 @@ class _Header extends HookConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final receiverPicture = ref.watch(
-      userMetadataFromDbNotifierProvider(receiverMasterPubkey).select((data) => data?.data.picture),
-    );
-
-    final receiverName = ref.watch(
-      userMetadataFromDbNotifierProvider(receiverMasterPubkey).select((data) => data?.data.name),
-    );
-
-    final receiverDisplayName = ref.watch(
-      userMetadataFromDbNotifierProvider(receiverMasterPubkey)
-          .select((data) => data?.data.displayName),
-    );
-
-    if (receiverName == null || receiverDisplayName == null) {
-      return const _HeaderSkeleton();
-    }
-
     return OneToOneMessagingHeader(
       conversationId: conversationId,
-      imageUrl: receiverPicture,
-      name: receiverDisplayName,
       receiverMasterPubkey: receiverMasterPubkey,
-      onTap: () => ProfileRoute(pubkey: receiverMasterPubkey).push<void>(context),
-      subtitle: Text(
-        prefixUsername(
-          context: context,
-          username: receiverName,
-        ),
-        style: context.theme.appTextThemes.caption.copyWith(
-          color: context.theme.appColors.quaternaryText,
-        ),
-      ),
-      onToggleMute: () {
-        ref.read(mutedConversationsProvider.notifier).toggleMutedMasterPubkey(receiverMasterPubkey);
-      },
-    );
-  }
-}
-
-class _HeaderSkeleton extends StatelessWidget {
-  const _HeaderSkeleton();
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsetsDirectional.fromSTEB(16.0.s, 8.0.s, 16.0.s, 12.0.s),
-      child: Row(
-        children: [
-          GestureDetector(
-            onTap: () => Navigator.pop(context),
-            child: Assets.svg.iconChatBack.icon(
-              size: 24.0.s,
-              flipForRtl: true,
-            ),
-          ),
-          SizedBox(width: 12.0.s),
-          ContainerSkeleton(
-            height: 36.0.s,
-            width: 36.0.s,
-            skeletonBaseColor: context.theme.appColors.onTerararyFill,
-          ),
-          SizedBox(width: 10.0.s),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                ContainerSkeleton(
-                  height: context.theme.appTextThemes.subtitle3.fontSize!.s,
-                  width: 120.0.s,
-                  skeletonBaseColor: context.theme.appColors.onTerararyFill,
-                ),
-                SizedBox(height: 4.0.s),
-                ContainerSkeleton(
-                  height: context.theme.appTextThemes.subtitle3.fontSize!.s,
-                  width: 150.0.s,
-                  skeletonBaseColor: context.theme.appColors.onTerararyFill,
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
